+++ crates/enclave-server/src/coco_aa/mod.rs
@@ -1,5 +1,3 @@
-pub mod handlers;
-
 use crate::key_manager;
 
 use anyhow::{anyhow, Result};
