+++ crates/enclave-server/src/coco_as/mod.rs
@@ -1,4 +1,3 @@
-pub mod handlers;
 pub mod into_original;
 pub mod policies;
