+++ crates/enclave-server/src/lib.rs
@@ -6,6 +6,5 @@ pub mod server;
 mod signing;
 pub mod snapshot; // pub for integration testing
 mod snapsync;
-mod tx_io;
 pub mod utils;
 pub mod api;
