--- crates/enclave-server/src/tx_io/mod.rs
@@ -1,1 +1,0 @@
-pub mod handlers;
